+++ src/indicators/IndicatorVariables.hpp
@@ -6,9 +6,10 @@ enum class IndicatorType {
     Clustered = 2
 };
 
-namespace IndicatorVars {
+namespace IndicatorVars { // im paranoid so everything gets a value
     inline cocos2d::CCDrawNode* triggerIndicatorDraw = nullptr;
     inline cocos2d::CCDrawNode* triggerExtraDraw = nullptr;
+    inline cocos2d::CCLayer* triggerIndicatorGroupLayer = nullptr;
     inline cocos2d::CCLayer* batchLayer = nullptr;
 
     inline bool disableMod = false;
@@ -18,13 +19,16 @@ namespace IndicatorVars {
     inline bool onlyTriggers = false;
     inline bool onlySpawn = false;
     inline bool spawnIndicators = false;
+    inline bool groupLabels = true;
     inline float maxDistance = 300.0f * 300.0f; // squaring cuz the distance function squares its results
     inline float thickness = 1.0f;
+    inline float spawnIndicatorThickness = 1.0f;
+    inline float groupLabelsSize = 1.0f;
     inline float clusterSize = 1.0f;
     inline int clusterMaxThreshold = 1;
     inline IndicatorType indicatorType = IndicatorType::Clustered;
-    inline cocos2d::ccColor4F extrasCol1 = cocos2d::ccc4f(1.0f, 1.0f, 1.0f, 1.0f);
-    inline cocos2d::ccColor4F extrasCol2 = cocos2d::ccc4f(1.0f, 1.0f, 1.0f, 1.0f);
+    inline cocos2d::ccColor4F extrasCol1 ={1.0f, 1.0f, 1.0f, 1.0f};
+    inline cocos2d::ccColor4F extrasCol2 = {1.0f, 1.0f, 1.0f, 1.0f};
     inline std::unordered_set<int> groupBlacklist = {
         1, 2, 3, 4, 5
     };